use case factored out into use case folder
