+++ src/main/java/UI/task_tree_UI/TaskTreeUI.java
@@ -1,7 +1,7 @@
 package UI.task_tree_UI;
 
 import UI.*;
-import UI.task_display_UI.TaskUI;
+import UI.task_UI.TaskUI;
 import use_cases.display_task_tree.*;
 
 import javax.swing.tree.DefaultMutableTreeNode;
