+++ src/main/java/use_cases/task_display_UI/TaskUiInputBoundary.java
@@ -1,4 +1,4 @@
-package UI.task_display_UI;
+package use_cases.task_display_UI;
 
 /**
  * An input boundary interface implemented by the TaskUiInteractor class.
