+++ src/main/java/use_cases/task_display_UI/TaskUiInteractor.java
@@ -1,4 +1,4 @@
-package UI.task_display_UI;
+package use_cases.task_display_UI;
 
 import entity_layer.Curriculum;
 import entity_layer.InMemoryUser;
