+++ src/main/java/use_cases/task_display_UI/TaskUiOutputBoundary.java
@@ -1,4 +1,4 @@
-package UI.task_display_UI;
+package use_cases.task_display_UI;
 
 /**
  * An output boundary interface implemented by the TaskUiModelPresenter class.
