+++ src/main/java/use_cases/task_display_UI/TaskUiViewInterface.java
@@ -1,4 +1,4 @@
-package UI.task_display_UI;
+package use_cases.task_display_UI;
 
 /**
  * An interface implemented by the TaskUI class.
