+++ src/test/java/use_cases/task_display_ui/TaskUiTestView.java
@@ -1,7 +1,7 @@
 package use_cases.task_display_ui;
 
-import UI.task_display_UI.TaskUiModel;
-import UI.task_display_UI.TaskUiViewInterface;
+import use_cases.task_display_UI.TaskUiModel;
+import use_cases.task_display_UI.TaskUiViewInterface;
 
 public class TaskUiTestView implements TaskUiViewInterface {
     private TaskUiModel taskUiModel;
